Fetch missing parents is now generic method.
