+++ src/senaite/sync/fetchstep.py
@@ -160,27 +160,3 @@ class FetchStep(SyncStep):
             return self.get_items("registry")
 
         return self.get_items("registry/{}".format(key))
-
-    def _fetch_missing_parents(self, item):
-        """
-        If data was fetched with portal type filter, this method will be used
-        to fill the missing parents for fetched objects.
-        :return:
-        """
-        if not self.content_types:
-            return
-
-        parent_path = item.get("parent_path")
-        # Skip if the parent is portal object
-        if len(parent_path.split("/")) < 3:
-            return
-        # Skip if already exists
-        if self.sh.find_unique("path", parent_path):
-            return
-
-        logger.info("Inserting missing parent: {}".format(parent_path))
-        parent = self.get_first_item(item.get("parent_url"))
-        par_dict = utils.get_soup_format(parent)
-        self.sh.insert(par_dict)
-        # Recursively import grand parents too
-        self._fetch_missing_parents(parent)
