+++ src/senaite/sync/syncstep.py
@@ -10,6 +10,7 @@ from BTrees.OOBTree import OOBTree
 from zope.annotation.interfaces import IAnnotations
 from senaite import api
 from senaite.sync import logger
+from senaite.sync import utils
 from senaite.sync.syncerror import SyncError
 
 SYNC_STORAGE = "senaite.sync"
@@ -161,3 +162,23 @@ class SyncStep:
         annotation = self.get_annotation()
         if annotation.get(SYNC_STORAGE) is not None:
             del annotation[SYNC_STORAGE]
+
+    def _fetch_missing_parents(self, item):
+        """
+        If data was fetched with portal type filter, this method will be used
+        to fill the missing parents for fetched objects.
+        :return:
+        """
+        parent_path = item.get("parent_path")
+        # Skip if the parent is portal object
+        if len(parent_path.split("/")) < 3:
+            return
+        # Skip if already exists
+        if self.sh.find_unique("path", parent_path):
+            return
+        logger.info("Inserting missing parent: {}".format(parent_path))
+        parent = self.get_first_item(item.get("parent_url"))
+        par_dict = utils.get_soup_format(parent)
+        self.sh.insert(par_dict)
+        # Recursively import grand parents too
+        self._fetch_missing_parents(parent)
